Sun, May 29, 2022, 4:55 AM +00:00

+++ calculator.c
@@ -4,11 +4,11 @@
 int main(void)
 {
     // Prompt user for x
-    int x = get_int("x: ");
+    long x = get_int("x: ");
 
     // Prompt user for y
-    int y = get_int("y: ");
+    long y = get_int("y: ");
 
     // Perform addition
-    printf("%i\n", x + y);
+    printf("%li\n", x + y);
 }
